01-image-generation/requirements.txt
Pillow==10.4.0
PyYAML==6.0.1
